Change parcel claim behavior to place items directly into inventory
- Modified RedeemParcel to use SummonItemIntoInventory instead of placing items on cursor
- Modified ReclaimAllParcels to use SummonItemIntoInventory
- Added inventory full checks to prevent items from dropping on ground
- If inventory is full, parcels remain in database for later claiming
- For non-stackable items, supports partial claiming with quantity updates
- Updated success messages to indicate items were placed in inventory

+++ Parcel-Manager/plugins/ParcelManager.pl
@@ -149,20 +149,65 @@ sub RedeemParcel {
         my $stacksize = $item_data ? int($item_data->{"stacksize"}) : 1;
         my $max_charges = $item_data ? int($item_data->{"maxcharges"}) : 0;
 
+        my $items_claimed = 0;
+        my $claim_success = 0;
+
         if ($max_charges > 0) {
             # Item with charges: quantity represents charges, summon one item with those charges
-            quest::summonitem($item_id, $quantity);
-            $message = "You have reclaimed $item_name with $quantity charges!";
+            my $item_obj = quest::summonitem($item_id, $quantity);
+            if (defined $item_obj) {
+                $claim_success = $client->SummonItemIntoInventory($item_obj);
+                if ($claim_success) {
+                    $items_claimed = 1;
+                    $message = "You have reclaimed $item_name with $quantity charges and it was placed in your inventory!";
+                } else {
+                    $client->Message(315, "Your inventory is full! Cannot claim $item_name. Please make space and try again.");
+                    $db->close();
+                    plugin::DisplayParcels();
+                    return;
+                }
+            }
         } elsif ($stacksize > 1) {
             # Stackable item: summon entire quantity at once
-            quest::summonitem($item_id, $quantity);
-            $message = "You have reclaimed $quantity x $item_name!";
+            my $item_obj = quest::summonitem($item_id, $quantity);
+            if (defined $item_obj) {
+                $claim_success = $client->SummonItemIntoInventory($item_obj);
+                if ($claim_success) {
+                    $items_claimed = $quantity;
+                    $message = "You have reclaimed $quantity x $item_name and it was placed in your inventory!";
+                } else {
+                    $client->Message(315, "Your inventory is full! Cannot claim $quantity x $item_name. Please make space and try again.");
+                    $db->close();
+                    plugin::DisplayParcels();
+                    return;
+                }
+            }
         } else {
             # Non-stackable: summon individual items
             for (my $i = 0; $i < $quantity; $i++) {
-                quest::summonitem($item_id);
+                my $item_obj = quest::summonitem($item_id);
+                if (defined $item_obj) {
+                    my $success = $client->SummonItemIntoInventory($item_obj);
+                    if ($success) {
+                        $items_claimed++;
+                    } else {
+                        # Inventory full - update parcel with remaining items
+                        my $remaining = $quantity - $items_claimed;
+                        if ($items_claimed > 0) {
+                            my $update_stmt = $db->prepare("UPDATE character_parcels SET quantity = ? WHERE id = ? AND char_id = ?");
+                            $update_stmt->execute($remaining, $parcel_id, $char_id);
+                            $update_stmt->close();
+                            $client->Message(315, "Claimed $items_claimed x $item_name before inventory became full. $remaining remaining in parcel.");
+                        } else {
+                            $client->Message(315, "Your inventory is full! Cannot claim $item_name. Please make space and try again.");
+                        }
+                        $db->close();
+                        plugin::DisplayParcels();
+                        return;
+                    }
+                }
             }
-            $message = "You have reclaimed $quantity x $item_name!";
+            $message = "You have reclaimed $quantity x $item_name and it was placed in your inventory!";
         }
     }
 
@@ -216,6 +261,7 @@ sub ReclaimAllParcels {
         my $parcel_id = $parcel->{id};
         my $item_id   = $parcel->{item_id};
         my $quantity  = $parcel->{quantity};
+        my $claim_success = 0;
 
         # Currency redemption (item_id 99990 is special currency parcel)
         if ($item_id == 99990) {
@@ -230,6 +276,7 @@ sub ReclaimAllParcels {
             quest::givecash($copper, $silver, $gold, $platinum);
 
             $parcels_claimed_data .= " | Money Claimed";
+            $claim_success = 1;
 
         } else {
             # Regular item redemption
@@ -246,33 +293,78 @@ sub ReclaimAllParcels {
 
             if ($max_charges > 0) {
                 # Item with charges: quantity represents charges, summon one item with those charges
-                quest::summonitem($item_id, $quantity);
-                $parcels_claimed_data .= " | $item_name ($quantity charges)";
+                my $item_obj = quest::summonitem($item_id, $quantity);
+                if (defined $item_obj) {
+                    $claim_success = $client->SummonItemIntoInventory($item_obj);
+                    if ($claim_success) {
+                        $parcels_claimed_data .= " | $item_name ($quantity charges)";
+                    } else {
+                        $client->Message(315, "Inventory full while claiming $item_name! Remaining parcels were not claimed.");
+                        $db->close();
+                        plugin::DisplayParcels();
+                        return;
+                    }
+                }
             } elsif ($stacksize > 1) {
                 # Stackable item: summon entire quantity at once
-                quest::summonitem($item_id, $quantity);
-                $parcels_claimed_data .= " | $item_name ($quantity)";
+                my $item_obj = quest::summonitem($item_id, $quantity);
+                if (defined $item_obj) {
+                    $claim_success = $client->SummonItemIntoInventory($item_obj);
+                    if ($claim_success) {
+                        $parcels_claimed_data .= " | $item_name ($quantity)";
+                    } else {
+                        $client->Message(315, "Inventory full while claiming $quantity x $item_name! Remaining parcels were not claimed.");
+                        $db->close();
+                        plugin::DisplayParcels();
+                        return;
+                    }
+                }
             } else {
                 # Non-stackable: summon individual items
+                my $items_claimed = 0;
                 for (my $i = 0; $i < $quantity; $i++) {
-                    quest::summonitem($item_id);
+                    my $item_obj = quest::summonitem($item_id);
+                    if (defined $item_obj) {
+                        my $success = $client->SummonItemIntoInventory($item_obj);
+                        if ($success) {
+                            $items_claimed++;
+                        } else {
+                            # Inventory full - update parcel with remaining items
+                            my $remaining = $quantity - $items_claimed;
+                            if ($items_claimed > 0) {
+                                my $update_stmt = $db->prepare("UPDATE character_parcels SET quantity = ? WHERE id = ? AND char_id = ?");
+                                $update_stmt->execute($remaining, $parcel_id, $char_id);
+                                $update_stmt->close();
+                                $client->Message(315, "Claimed $items_claimed x $item_name before inventory became full. $remaining remaining in this parcel. Remaining parcels were not claimed.");
+                            } else {
+                                $client->Message(315, "Inventory full while claiming $item_name! Remaining parcels were not claimed.");
+                            }
+                            $db->close();
+                            plugin::DisplayParcels();
+                            return;
+                        }
+                    }
                 }
-                $parcels_claimed_data .= " | $item_name ($quantity)";
+                if ($items_claimed == $quantity) {
+                    $claim_success = 1;
+                    $parcels_claimed_data .= " | $item_name ($quantity)";
+                }
             }
         }
 
-        $total_claimed++;
-
-        # Remove parcel from database
-        my $del_stmt = $db->prepare("DELETE FROM character_parcels WHERE id = ? AND char_id = ?");
-        $del_stmt->execute($parcel_id, $char_id);
-        $del_stmt->close();
+        # Only remove parcel from database if successfully claimed
+        if ($claim_success) {
+            $total_claimed++;
+            my $del_stmt = $db->prepare("DELETE FROM character_parcels WHERE id = ? AND char_id = ?");
+            $del_stmt->execute($parcel_id, $char_id);
+            $del_stmt->close();
+        }
     }
     $db->close();
 
     # Send confirmation message and cleanup
     if ($total_claimed > 0) {
-        $client->Message(315, "Successfully claimed $total_claimed parcel(s)!");
+        $client->Message(315, "Successfully claimed $total_claimed parcel(s) and placed them in your inventory!");
         quest::delete_data("PARCELS_$char_id"); # Clear QGlobal since all parcels claimed
     } else {
         $client->Message(315, "No parcels were found to claim.");
